requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-cors-headers==2.1.0
django-react-templatetags==300
djangorestframework==3.7.3
enum34==1.1.6
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
npm==0.1.1
numpy==1.13.3
oauth2client==4.1.2
optional-django==0.1.0
pandas==0.21.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygsheets==1.1.3
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
